refactor: ssr과 헤더 분리
