+++ pages/upload/[id].tsx
@@ -4,7 +4,6 @@ import { Loading } from "components/custom";
 import { useQuery } from "react-query";
 import Direct from "./direct";
 import Embed from "./embed";
-import { NextPageWithLayout } from "types/common";
 import { GetServerSideProps } from "next";
 
 export const getServerSideProps: GetServerSideProps = async (context) => {
