+++ styles/search/cate.module.css
@@ -1,105 +1,3 @@
-@value cateWhole, cateDog, cateCat, cateLizard, cateBird, cateRabbit, cateHamster, cateETC from 'styles/variables.css';
-
-.header {
-  max-width: 480px;
-  overflow: hidden;
-  display: flex;
-  position: fixed;
-  padding: 15px 10px;
-  width: 100%;
-  top: 0;
-  left: 50%;
-  transform: translateX(-50%);
-  height: 100px;
-  z-index: 1000;
-  border-bottom-left-radius: 8px;
-  border-bottom-right-radius: 8px;
-}
-
-.cate_name {
-  position: absolute;
-  bottom: 15px;
-  left: 40px;
-  color: white;
-  font-weight: 700;
-  font-size: 25px;
-}
-.cate_whole {
-  background-color: cateWhole;
-}
-
-.cate_dog {
-  background-color: cateDog;
-}
-
-.cate_cat {
-  background-color: cateCat;
-}
-
-.cate_lizard {
-  background-color: cateLizard;
-}
-
-.cate_bird {
-  background-color: cateBird;
-}
-
-.cate_rabbit {
-  background-color: cateRabbit;
-}
-
-.cate_hamster {
-  background-color: cateHamster;
-}
-
-.cate_etc {
-  background-color: cateETC;
-}
-
-.dog_container {
-  position: absolute;
-  right: 30px;
-  overflow: hidden;
-}
-
-.cat_container {
-  position: absolute;
-  right: 50px;
-  overflow: hidden;
-}
-
-.lizard_container {
-  position: absolute;
-  right: 0px;
-  bottom: -4px;
-  overflow: hidden;
-}
-
-.bird_container {
-  position: absolute;
-  right: 0px;
-  overflow: hidden;
-}
-
-.rabbit_container {
-  position: absolute;
-  right: 50px;
-  overflow: hidden;
-}
-
-.hamster_container {
-  position: absolute;
-  right: 40px;
-  bottom: -4px;
-  overflow: hidden;
-}
-
-.etc_container {
-  position: absolute;
-  right: 40px;
-  bottom: -4px;
-  overflow: hidden;
-}
 /*select box*/
 .select_container {
   max-width: 480px;
